Move pkg/util/interrupt to apimachinery
This package is a helper to assist in doing guaranteed cleanup on signals,
primarily used by kubectl, but also by other one off processes that might
run in pods. Move it out of k8s.io/kubernetes first, before potentially
moving it to a library package or other.
