+++ staging/src/BUILD
@@ -44,6 +44,7 @@ filegroup(
         "//staging/src/k8s.io/apimachinery/pkg/util/framer:all-srcs",
         "//staging/src/k8s.io/apimachinery/pkg/util/httpstream:all-srcs",
         "//staging/src/k8s.io/apimachinery/pkg/util/initialization:all-srcs",
+        "//staging/src/k8s.io/apimachinery/pkg/util/interrupt:all-srcs",
         "//staging/src/k8s.io/apimachinery/pkg/util/intstr:all-srcs",
         "//staging/src/k8s.io/apimachinery/pkg/util/json:all-srcs",
         "//staging/src/k8s.io/apimachinery/pkg/util/jsonmergepatch:all-srcs",
