+++ staging/src/k8s.io/apimachinery/pkg/util/interrupt/BUILD
@@ -1,14 +1,11 @@
-package(default_visibility = ["//visibility:public"])
-
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-)
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
     srcs = ["interrupt.go"],
-    importpath = "k8s.io/kubernetes/pkg/util/interrupt",
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/apimachinery/pkg/util/interrupt",
+    importpath = "k8s.io/apimachinery/pkg/util/interrupt",
+    visibility = ["//visibility:public"],
 )
 
 filegroup(
@@ -22,4 +19,5 @@ filegroup(
     name = "all-srcs",
     srcs = [":package-srcs"],
     tags = ["automanaged"],
+    visibility = ["//visibility:public"],
 )
